backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
mediapipe==0.10.9
opencv-python==4.9.0.80
numpy==1.26.3
gunicorn==21.2.0
